parte do ajax para gestão de projetos

+++ app/Http/Controllers/ProjetoController.php
@@ -5,11 +5,13 @@ namespace App\Http\Controllers;
 use App\Models\Projeto;
 use Illuminate\Http\Request;
 use App\Http\Middleware\IsAdminMiddleware;
+use App\Models\User;
 
 
 class ProjetoController extends Controller
 {
-    public function __construct() {
+    public function __construct(Projeto $projeto) {
+        $this->projeto = $projeto;
         //$this->middleware(IsAdminMiddleware::class);
     }
 
@@ -21,12 +23,16 @@ class ProjetoController extends Controller
     public function index()
     {
         //throw new \Exception('Erro ao adicionar projeto!');
+        $usuariosComuns = User::select('id', 'name')->where('admin', '=', false)->orderBy('name')->get();
+
         if (!auth()->user()->admin) {
-            $projetos = Projeto::where('user_id', '=', auth()->user()->id)->paginate(10);
+            $projetos = $this->projeto->where('user_id', '=', auth()->user()->id)->orderBy('nome')->paginate(10);
         } else {
-            $projetos = Projeto::paginate(10);
+            $projetos = $this->projeto->orderBy('nome')->paginate(10);
         }
-        return view('app.projeto.index', ['projetos' => $projetos]);
+        $data = ['projetos' => $projetos, 'usuarios_comuns' => $usuariosComuns];
+
+        return view('app.projeto.index', $data);
     }
 
     /**
@@ -36,7 +42,7 @@ class ProjetoController extends Controller
      */
     public function create()
     {
-        return View('app.projeto.create');
+        //return View('app.projeto.create');
     }
 
     /**
@@ -47,19 +53,27 @@ class ProjetoController extends Controller
      */
     public function store(Request $request)
     {
-        $projeto = Projeto::create($request->all());
-        return redirect()->route('projeto.show', ['projeto' => $projeto->id]);
+        $request->validate($this->projeto->rules(), $this->projeto->feedback());
+        $projeto = $this->projeto->create($request->all());
+        $resultado = ['success' => true, 'data' => $projeto];
+
+        return response()->json($projeto, 201);
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Projeto  $projeto
+     * @param  Integer  $id
      * @return \Illuminate\Http\Response
      */
-    public function show(Projeto $projeto)
+    public function show($id)
     {
-        return view('app.projeto.show', ['projeto' => $projeto]);
+        $projeto = $this->projeto->find($id);
+
+        if ($projeto === null) {
+            return response()->json(['erro' => 'Projeto não encontrado.'], 404);
+        }
+        return response()->json($projeto, 200);
     }
 
     /**
@@ -70,35 +84,51 @@ class ProjetoController extends Controller
      */
     public function edit(Projeto $projeto)
     {
-        return view('app.projeto.edit', ['projeto' => $projeto]);
+        //return view('app.projeto.edit', ['projeto' => $projeto]);
     }
 
     /**
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Projeto  $projeto
+     * @param  Integer  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Projeto $projeto)
+    public function update(Request $request, $id)
     {
+        $projeto = $this->projeto->find($id);
+
+        if ($projeto === null) {
+            return response()->json(['erro' => 'Não foi possível realizar a atualização. Projeto não existe.'], 404);
+        }
+
+        $request->validate($this->projeto->rules(), $this->projeto->feedback());
         $projeto->update($request->all());
-        return redirect()->route('projeto.show', ['projeto' => $projeto->id]);
+
+        return response()->json($projeto, 200);
+        //return redirect()->route('projeto.show', ['projeto' => $projeto->id]);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Projeto  $projeto
+     * @param  \Integer  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Projeto $projeto)
+    public function destroy($id)
     {
+        $projeto = $this->projeto->find($id);
+        if ($projeto === null) {
+            return response()->json(['erro' => 'Impossível remover projeto. Projeto informado não existe.'], 404);
+        }
         $projeto->delete();
-        return redirect()->route('projeto.index');
+
+        return response()->json(['message' => 'Projeto removido com sucesso']);
     }
 
-    public function meusProjetos() {
-        dd('teste P ok!!!');
+    public function listaSimplesDeUsuarios() {
+        // TODO : Rever a função e necessidade de implementar.
+        $usuarios = User::select('id', 'name')->where('admin', '=', false)->get();
+        return json_encode($usuarios);
     }
 }
